Fix k8s-service selector issue
When a jointinferenceservice is created, a k8s service is created for
big worker, so the big-worker's lables need to be unique.

This can be done by adding a new label key 'worker-type'.

Same as federatedlearningjob.
